Fix Aether parachutes causing permanent crash 🪂
After the player falls out of the Aether dimension with a
parachute equipped, when re-entering the Aether the parachute
is still there and due to being marked as being mounted with the
player, causes a crash preventing the Aether from being entered
again.

This commit causes parachutes to despawn when they get below y=0,
and also lowers the height in which the player will 'fall out of'
the Aether. (gives the parachute more time to despawn and also it's
really annoying flying in creative falling at y<-2)

Code from Departure, I'm just the git gremlin

Signed-off-by: Richard Rogalski <rrogalski@firemail.cc>
Co-authored-by: Wojak <102751411+AetherLegacyPro@users.noreply.github.com>